requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.8
packaging==23.2
Pillow==10.1.0
pytesseract==0.3.6
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
